requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers[torch]>=0.23.0
transformers>=4.35.0
gradio>=4.0.0
numpy>=1.21.0
matplotlib>=3.5.0
opencv-python>=4.5.0
tqdm>=4.64.0
rembg[gpu]>=2.0.50
onnxruntime-gpu>=1.16.0
PyYAML==5.1
scipy
scikit-image
ftfy
regex
accelerate
Pillow>=7.1,  
matplotlib
pycocotools>=2.0.2
termcolor>=1.1
yacs>=0.1.8
tabulate
cloudpickle
tqdm>4.29.0
tensorboard
fvcore>=0.1.5,<0.1.6
iopath>=0.1.7,<0.1.10
dataclasses; python_version<'3.7'
omegaconf>=2.1,<2.4
hydra-core>=1.1
black
packaging
av